requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.4
django-bootstrap4==2.3.1
django-stdimage==5.1.1
gunicorn==20.0.4
importlib-metadata==2.1.1
Pillow==8.0.1
progressbar2==3.53.1
psycopg2-binary==2.8.6
python-utils==2.4.0
pytz==2020.4
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
static3==0.7.0
zipp==3.4.0
